move bill of sale schema
